don't let vert.x RoutingContext decorator to move failures to another thread
I'm not sure why it does this, but the original routing context doesn't force the issue, so overload the decorator and call the original logic - it allows for simpler failure handling in the handler implementation.

If the handler implementation wants to move the failure to a worker thread, they are welcome to do so in all the ways that Vert.x offers, and also by sub-classing the request object and doing their own fail() implementation - as Irked was designed for.