Lock Version 1 with support for Vert.x ~> 3.8 only.
Vert.x 3.9 is not API backward compatible with Vert.x 3.8 as it changed
some methods return type. If Irked calls these methods and compiles with
Vert.x 3.9, but then run with Vert.x 3.8, we get NoSuchMethod
exceptions.

We will issue Version 2 that will be compatible with Vert.x 3.9 and
later Version 3 for compatibility with the completely new Vert.x 4 API.