remove the service loader for authorization tokens due to security concerns.
The service loader is a problem in the case of fuzzing attacks as it
will hit the disk every time it is asked to load an unknown
implementation and will not cache negative results (not that we expect
it to waste memory on that). By moving to a static list of
implementations we lose the ability for users to extend the
implementation with non-standard authorization tokens - but hopefully
this is not a requested feature.